ffmpeg fix
Hi Thaddeus,
I managed to make the processor work, but it has some caveats, so here's a quick summary:
I tried to test it with my own video and I had the same problem as You had, and the root cause seems to be that OpenCV is built without any proper Video I/O packages, for example on linux only v4l/v4l2 was available when building OpenCV which is probably enough only for a very limited number of video formats.
For most video formats OpenCV should probably be built with FFMPEG enabled, so it has the proper codecs and stream processing libs. I enabled FFMPEG for the OpenCV build which looks for the system installed FFMPEG packages for the build so I had to install the following packages (I'm working on ubuntu 22.04):
sudo apt install ffmpeg libavcodec-dev libavformat-dev libavutil-dev libswscale-dev
We currently link OpenCV to the libminifi-opencv extension statically, and for the FFMPEG to work with our library it needs to be linked to this extension as well. Unfortunately we cannot statically link the system libraries to our extension due to different compile options, so the ffmpeg needs to be built as a separate OpenCV plugin shared library which could be loaded dynamically. To do this I used the following patch in MiNiFi:
diff --git a/cmake/BundledOpenCV.cmake b/cmake/BundledOpenCV.cmake
index ee4a7ff94..56de9ffd1 100644
--- a/cmake/BundledOpenCV.cmake
+++ b/cmake/BundledOpenCV.cmake
@@ -40,6 +40,7 @@ function(use_bundled_opencv SOURCE_DIR BINARY_DIR)
     set(OPENCV_BYPRODUCT_DIR "${CMAKE_CURRENT_BINARY_DIR}/thirdparty/opencv-install")
     set(BYPRODUCTS
         "${OPENCV_BYPRODUCT_DIR}/${LIBDIR}/${PREFIX}opencv_flann${SUFFIX}"
+        "${OPENCV_BYPRODUCT_DIR}/${LIBDIR}/${PREFIX}opencv_dnn${SUFFIX}"
         "${OPENCV_BYPRODUCT_DIR}/${LIBDIR}/${PREFIX}opencv_objdetect${SUFFIX}"
         "${OPENCV_BYPRODUCT_DIR}/${LIBDIR}/${PREFIX}opencv_core${SUFFIX}"
         "${OPENCV_BYPRODUCT_DIR}/${LIBDIR}/${PREFIX}opencv_gapi${SUFFIX}"
@@ -68,6 +69,7 @@ function(use_bundled_opencv SOURCE_DIR BINARY_DIR)
             "-DBUILD_opencv_calib3d=ON"
             "-DBUILD_opencv_core=ON"
             "-DBUILD_opencv_dnn=OFF"
+            "-DOPENCV_PLUGIN_NAME=opencv_videoio_ffmpeg"
             "-DBUILD_opencv_features2d=ON"
             "-DBUILD_opencv_flann=ON"
             "-DBUILD_opencv_gapi=ON"
@@ -80,13 +82,14 @@ function(use_bundled_opencv SOURCE_DIR BINARY_DIR)
             "-DBUILD_opencv_stitching=ON"
             "-DBUILD_opencv_video=ON"
             "-DBUILD_opencv_videoio=ON"
+            "-DVIDEOIO_PLUGIN_LIST=ffmpeg"
             "-DBUILD_JAVA=OFF"
             "-DBUILD_FAT_JAVA_LIB=OFF"
             "-DBUILD_PNG=ON"
             "-DBUILD_JPEG=ON"
             "-DBUILD_OPENJPEG=OFF"
             "-DWITH_1394=OFF"
-            "-DWITH_FFMPEG=OFF"
+            "-DWITH_FFMPEG=ON"
             "-DWITH_GSTREAMER=OFF"
             "-DWITH_GTK=OFF"
             "-DWITH_IPP=OFF"
After the build is done the OpenCV build will create the plugin at ${minifi_root}/build/extensions/opencv/thirdparty/opencv-install/libopencv_videoio_ffmpeg.so which needs to be copied in the minifi binary directory, for the OpenCV extension to be able to load it dynamically.
After all this I could make the test pass capturing the video frames successfully. I know this is not ideal currently and will need to be fixed. Maybe in the future we can either build our own FFMPEG library and link it statically to the OpenCV extension so it has all the needed features in one place, or change the extension to use system installed OpenCV and FFMPEG or any video I/O packages and load all of them dynamically.